5.5 Blog list frontend, step5
MODIFIED creating notes form to show only when appropriate. CREATED a
component "Togglable" that handles the logic for showing and hiding
the component wrapped in it. The useRef, useImperativeHandle,
forwardRef hooks were used to pass the toggling function to the parent
component. It was passed upwards to enable hiding of the blog form upon
successful creation of blogs.